sockets: splitting connection socket out of listen_socket_impl
